moved test directory

+++ test/test_crc.cpp
@@ -4,7 +4,7 @@
 #include <iostream>
 
 #include "catch.hpp"
-#include "datagroups.h"
+#include "../src/common.h"
 
 using namespace std;
 
@@ -14,7 +14,7 @@ TEST_CASE( "CRC encoding tests", "[msc]" ) {
 		string data("123456789");
 		vector<unsigned char> bytes;
 		copy(data.begin(), data.end(), back_inserter(bytes));
-        unsigned int crc = calculate_crc(bytes);
+        unsigned short crc = calculate_crc(bytes);
         REQUIRE(crc == 0x906e);
 	}
